feat(tm2): add transfer event for bank ops
