+++ gnovm/stdlibs/chain/amino_package.go
@@ -18,5 +18,4 @@ var Package = amino.RegisterPackage(amino.NewPackage(
 		Event{},
 		StorageDepositEvent{},
 		StorageUnlockEvent{},
-		TransferEvent{},
 	))
