+++ gnovm/stdlibs/chain/emit_event.go
@@ -8,7 +8,6 @@ import (
 
 	gno "github.com/gnolang/gno/gnovm/pkg/gnolang"
 	"github.com/gnolang/gno/gnovm/stdlibs/internal/execctx"
-	"github.com/gnolang/gno/tm2/pkg/crypto"
 	"github.com/gnolang/gno/tm2/pkg/std"
 )
 
@@ -124,12 +123,3 @@ type StorageUnlockEvent struct {
 }
 
 func (e StorageUnlockEvent) AssertABCIEvent() {}
-
-// TransferEvent is emitted when coins are transferred between accounts.
-type TransferEvent struct {
-	From   crypto.Address `json:"from"`
-	To     crypto.Address `json:"to"`
-	Amount std.Coins      `json:"amount"`
-}
-
-func (TransferEvent) AssertABCIEvent() {}
